pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "governed-analytics-copilot"
version = "0.1.0"
description = "A governed NL-to-SQL analytics copilot with a semantic layer"
requires-python = ">=3.10"
dependencies = [
    # ── Web / API ────────────────────────────────────
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    # ── Database ─────────────────────────────────────
    "sqlalchemy>=2,<3",
    "psycopg2-binary>=2.9,<3",
    # ── dbt ──────────────────────────────────────────
    "dbt-core>=1.7,<2",
    "dbt-postgres>=1.7,<2",
    # ── Data quality ─────────────────────────────────
    "great-expectations>=0.18,<1",
    # ── UI ───────────────────────────────────────────
    "streamlit>=1.32,<2",
    # ── LLM (optional) ──────────────────────────────
    "openai>=1.12,<2",
    "anthropic>=0.21,<1",
    # ── Utilities ────────────────────────────────────
    "python-dotenv>=1,<2",
    "pyyaml>=6,<7",
    "httpx>=0.27,<1",
    "sqlparse>=0.5,<1",
    "tabulate>=0.9,<1",
    "Faker>=24,<30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.3,<1",
    "pre-commit>=3,<4",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
